添加运行时参数存储模型；添加模型版本新增服务

+++ src/Dist.Product.Dme/Dist.Dme.RuleSteps/BaseRuleStepMeta.cs
@@ -120,7 +120,7 @@ namespace Dist.Dme.RuleSteps
                     VersionId = this.step.VersionId,
                     AttributeCode = item.Key,
                     AttributeValue = item.Value.Value,
-                    AttributeType = item.Value.Type
+                    AttributeType = (int)item.Value.Type
                 };
 
                 if (nameof(EnumValueMetaType.TYPE_FEATURECLASS).Equals(item.Value.DataTypeCode)
